+++ packages/bundler/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Dropped support for node v6. This is a soft break, as we aren't
   making any changes that are known to break node v6, but we're no longer testing against it. See our [node version support policy](https://www.polymer-project.org/2.0/docs/tools/node-support)
   for details.
+* Added support for `import.meta.url` preservation in bundled modules.
 <!-- Add new, unreleased changes here. -->
 
 ## 4.0.0-pre.6 - 2018-04-25